Move away from boolean use indicating requirement.
The way to specify how if an option was required was to pass a
boolean flag in the constructor. This is not clear, and not
extensible. Instead use an enumeration to provide the same
functionality, for example:
```java
new BasicOption ("long", "l", true) // not clear
new BasicOption ("long", "l", RequirePresence)
```

The requirements enum is located under `Option.Requirement`. This
enables us to more easily implement throwing errors when options are
required. It is also easier to indicate whether an option requires a
value on top of its presence.